Each minute is 1 second, each hour is 1 minute
To test the songs of the clock here is a quick time version. Honestly you should let your ears rest from the passive buzzer for a minute anyways. Please silence you buzzer by either covering it or using resistors or some other method. Doesn't seem loud at first, but not good for the ears. 1 "minute" is set to delay(1000); If you want it more frequent, then you can change the duration length of the minute.